LP-171 refactor : favorite 패키지 이름 변경
