--- src/main/java/com/linkmoa/source/domain/Favorite/constant/ItemType.java
@@ -1,5 +1,0 @@
-package com.linkmoa.source.domain.Favorite.constant;
-
-public enum ItemType {
-	SITE, DIRECTORY
-}
