+++ src/main/java/com/linkmoa/source/domain/directory/repository/DirectoryRepositoryImpl.java
@@ -1,16 +1,16 @@
 package com.linkmoa.source.domain.directory.repository;
 
-import static com.linkmoa.source.domain.Favorite.entity.QFavorite.*;
 import static com.linkmoa.source.domain.directory.entity.QDirectory.*;
+import static com.linkmoa.source.domain.favorite.entity.QFavorite.*;
 import static com.linkmoa.source.domain.site.entity.QSite.*;
 
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import com.linkmoa.source.domain.Favorite.constant.ItemType;
 import com.linkmoa.source.domain.directory.dto.response.DirectoryDetailResponse;
 import com.linkmoa.source.domain.directory.entity.Directory;
+import com.linkmoa.source.domain.favorite.constant.ItemType;
 import com.querydsl.core.types.Projections;
 import com.querydsl.core.types.dsl.Expressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
