+++ src/main/java/com/linkmoa/source/domain/favorite/constant/ItemType.java
@@ -1,0 +1,5 @@
+package com.linkmoa.source.domain.favorite.constant;
+
+public enum ItemType {
+	SITE, DIRECTORY
+}
