+++ src/main/java/com/linkmoa/source/domain/favorite/repository/FavoriteRepository.java
@@ -1,11 +1,11 @@
-package com.linkmoa.source.domain.Favorite.repository;
+package com.linkmoa.source.domain.favorite.repository;
 
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.linkmoa.source.domain.Favorite.constant.ItemType;
-import com.linkmoa.source.domain.Favorite.entity.Favorite;
+import com.linkmoa.source.domain.favorite.constant.ItemType;
+import com.linkmoa.source.domain.favorite.entity.Favorite;
 import com.linkmoa.source.domain.member.entity.Member;
 
 public interface FavoriteRepository extends JpaRepository<Favorite, Long>, FavoriteRepositoryCustom {
